{{cookiecutter.hyphenated_name}}/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
appdirs==1.4.4            # via black
black==20.8b1             # via -r requirements.in
certifi==2020.6.20        # via requests
chardet==3.0.4            # via requests
click==7.1.2              # via black
deprecated==1.2.10        # via pygithub
environs==8.0.0           # via -r requirements.in
flake8==3.8.3             # via -r requirements.in
idna==2.10                # via requests
importlib-metadata==2.0.0  # via flake8
isort==5.5.3              # via -r requirements.in
marshmallow==3.8.0        # via environs
mccabe==0.6.1             # via flake8
mypy-extensions==0.4.3    # via black
pathspec==0.8.0           # via black
pycodestyle==2.6.0        # via flake8
pyflakes==2.2.0           # via flake8
pygithub==1.53            # via -r requirements.in
pyjwt==1.7.1              # via pygithub
python-dotenv==0.14.0     # via environs
pyupgrade==2.7.2          # via -r requirements.in
regex==2020.7.14          # via black
requests==2.24.0          # via pygithub
tokenize-rt==4.0.0        # via pyupgrade
toml==0.10.1              # via black
typed-ast==1.4.1          # via black
typing-extensions==3.7.4.3  # via black
urllib3==1.25.10          # via requests
wrapt==1.12.1             # via deprecated
zipp==3.2.0               # via importlib-metadata
